first task

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -13,17 +13,17 @@ int main(void)
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
-	int last_digit;
+	float last_digit;
 	last_digit = n % 10;
 
 if(last_digit > 5)
 {
-	printf("Last digit of %d is %d and is greater than 5\n ",n, last_digit);
+	printf("Last digit of %d is %1.0f and is greater than 5\n ",n, last_digit);
 }else if(last_digit == 0)
 {
-	printf("Last digit of %d is %d and is zero\n", n, last_digit);
+	printf("Last digit of %d is %1.0f and is zero\n", n, last_digit);
 }else if(last_digit < 6)
 {
-	printf("Last digit of %d is %d and is less than 6 and not zero\n", n, last_digit);
+	printf("Last digit of %d is %1.0f and is less than 6 and not zero\n", n, last_digit);
 }
 }
